+++ it/google-cloud-platform/src/test/java/org/apache/beam/it/gcp/dataflow/ClassicTemplateClientTest.java
@@ -79,7 +79,7 @@ public final class ClassicTemplateClientTest {
   @Captor private ArgumentCaptor<CreateJobFromTemplateRequest> requestCaptor;
 
   @Test
-  public void testCreateWithCredentials() throws IOException {
+  public void testCreateWithCredentials() {
     Credentials credentials = mock(Credentials.class);
     ClassicTemplateClient.builder(credentials).build();
     // Lack of exception is all we really can test