+++ sdks/python/apache_beam/io/gcp/bigtableio_it_test.py
@@ -50,6 +50,7 @@ except ImportError as e:
 
 
 @pytest.mark.uses_gcp_java_expansion_service
+@pytest.mark.uses_transform_service
 @unittest.skipUnless(
     os.environ.get('EXPANSION_PORT'),
     "EXPANSION_PORT environment var is not provided.")
@@ -142,6 +143,7 @@ class TestReadFromBigTableIT(unittest.TestCase):
 
 
 @pytest.mark.uses_gcp_java_expansion_service
+@pytest.mark.uses_transform_service
 @unittest.skipUnless(
     os.environ.get('EXPANSION_PORT'),
     "EXPANSION_PORT environment var is not provided.")